Rename `components` page to `component-list`
