+++ src/app/pages/component-list/component-list.html
@@ -1,0 +1,8 @@
+<div *ngFor="let component of componentItems">
+  <img src="../../../assets/img/components/{{component.src}}.svg" [alt]="component.name">
+  {{component.name}}
+</div>
+
+
+
+<!--<doc-viewer documentUrl="/assets/documents/README.html"></doc-viewer>-->
