+++ src/app/pages/component-list/component-list.spec.ts
