+++ src/app/pages/component-list/index.ts
@@ -1,0 +1,1 @@
+export * from './component-list';
