--- src/app/pages/components/components.html
@@ -1,5 +1,0 @@
-<div *ngFor="let component of componentItems">
-  <img src="../../../assets/img/components/{{component.src}}.svg" [alt]="component.name" />
-  {{component.name}}
-</div>
-<doc-viewer documentUrl="/assets/documents/README.html"></doc-viewer>
