+++ src/app/routes.ts
@@ -1,11 +1,11 @@
 import { ModuleWithProviders } from '@angular/core';
 import { Homepage } from './pages/homepage';
-import { ComponentsList } from './pages/components';
+import { ComponentList } from './pages/component-list';
 import { Routes, RouterModule } from '@angular/router';
 
 const MATERIAL_DOCS_ROUTES: Routes = [
   { path: '', component: Homepage, pathMatch: 'full' },
-  { path: 'components', component:  ComponentsList },
+  { path: 'components', component:  ComponentList },
 ];
 
 export const routing: ModuleWithProviders = RouterModule.forRoot(MATERIAL_DOCS_ROUTES);
